requirements.txt
paramiko==3.4.0
openai==1.12.0
pymongo==4.6.1
python-dotenv==1.0.0
torch==2.2.0
numpy==1.26.3
langfuse==2.7.0
google-cloud-storage==2.18.0
google-auth==2.30.0 
assemblyai==0.40.2
fastapi[standard]
groq
langfuse
# OpenTelemetry packages (compatible with Python 3.12)
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-instrumentation-redis>=0.41b0
opentelemetry-instrumentation-pymongo>=0.41b0
opentelemetry-exporter-jaeger>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-logging>=0.41b0

langchain
langchain-groq==0.3.1
langchain-openai>=0.2.0
langgraph>=0.2.0
python-jose[cryptography]>=3.3.0
passlib==1.7.4
pydantic[email]>=2.3.0
bcrypt==4.0.1
redis>=4.2.0,<5.0.0
aioredis==2.0.1
fastapi-limiter==0.1.5
pymysql==1.1.1 
google-genai==1.16.1
google-generativeai==0.8.5
python-dotenv>=1.0.0
httpx>=0.25.0
PyJWT>=2.8.0

# Audio Monitor Service Dependencies
cryptography>=41.0.0
motor>=3.3.0
assemblyai>=0.17.0
openai>=1.3.0
google-cloud-storage>=2.10.0
paramiko>=3.3.0